138 생산 보완

+++ src/main/java/com/itwillbs/dto/WorkOrderMaterialDTO.java
@@ -1,5 +1,7 @@
 package com.itwillbs.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 /**
@@ -11,5 +13,5 @@ import lombok.Data;
 public class WorkOrderMaterialDTO {
     private String workOrderId;   // 작업지시 ID (FK)
     private String materialId;    // 자재 코드
-    private double requiredQty;   // 총 소요량(소수점 포함 가능)
+    private BigDecimal requiredQty;  // 총 소요량(소수점 포함 가능)
 }
